wip(setup): Initialize npm into project folder
We need npm in order to use webpack, so we initialize it using the command `npm init`.
Optionally you can add the flag `-y` to not answer the questions and fill/update them
later.

This will produce a file called `package.json`, used for `npm` to figure out the setup.
I encourage you to properly fill the `package.json` using a good text editor.
 Committer: Israel Laguan <israel@desktop>

 On branch one-js
 Changes to be committed:
	new file:   package.json